% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmextra-package.R
\docType{data}
\name{acacia}
\alias{acacia}
\title{Native and introduced range of Acacia cyclops}
\format{
A list of two
\describe{
  \item{south_australia}{\code{SpatialPolygonsDataFrame} of the native range (reference area).}
  \item{south_africa}{\code{SpatialPolygonsDataFrame} of the introduced range (target area).}
}
}
\source{
Sample data from the ExDet tutorial available from https://www.climond.org/ExDet.aspx.
}
\description{
Spatial extents of the native and introduced ranges of A. cyclops, a small shrub species.
}
\details{
#' Mesgaran MB, Cousens RD, Webber BL (2014). Here be dragons: a tool for quantifying novelty due to covariate range and correlation change when projecting species distribution models. Diversity & Distributions, 20: 1147-1159, DOI: \href{https://onlinelibrary.wiley.com/doi/full/10.1111/j.1472-4642.2011.00811.x}{10.1111/j.1472-4642.2011.00811.x}

Webber BL, Yates, CJ, Le Maitre DC, Scott JK, Kriticos DJ, Ota N, McNeill A, Le Roux JJ, Midgley GF (2011). Modelling horses for novel climate courses: Insights from projecting potential distributions of native and alien Australian acacias with correlative and mechanistic models. Diversity and Distributions, 17: 978–1000. DOI: \href{https://onlinelibrary.wiley.com/doi/full/10.1111/ddi.12209}{10.1111/ddi.12209}
}
\keyword{datasets}
