% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_crs}
\alias{check_crs}
\title{---------------------------------------------
---------------------------------------------}
\usage{
check_crs(coordinate.system)
}
\description{
---------------------------------------------
---------------------------------------------
}
