% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse.vars}
\alias{collapse.vars}
\title{---------------------------------------------
---------------------------------------------}
\usage{
collapse.vars(l)
}
\description{
---------------------------------------------
---------------------------------------------
}
