% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_extrapolation.R
\name{compute_extrapolation}
\alias{compute_extrapolation}
\title{Quantify extrapolation in multivariate environmental space}
\usage{
compute_extrapolation(segments, covariate.names, prediction.grid,
  coordinate.system, print.summary = TRUE, print.precision = 2,
  save.summary = FALSE, resolution = NULL)
}
\arguments{
\item{segments}{Segment data.frame (i.e. surveyed transects divivided into segments for analysis). This is the reference dataset used for model building and calibration. This must contain one column for each of the covariates in `covariate.names`.}

\item{covariate.names}{Character string. Names of the covariates of interest.}

\item{prediction.grid}{Prediction data.frame. This contains both the geographic coordinates (`x`, `y`) and the covariate values associated with the target locations for which predictions are desired. Typically, these locations are taken as the centroids of the grid cells in a spatial prediction grid/raster.}

\item{coordinate.system}{Projected coordinate system relevant to the study location. Can be either a character string or an object of class \code{\link[sp]{CRS}}.}

\item{print.summary}{Logical, defaults to TRUE. Outputs a summary of the results to the console}

\item{print.precision}{Integer. Number of significant figures to be used when printing the summary. Default value of 2.}

\item{save.summary}{Logical. Adds summary statistics to the output list. Defaults to FALSE.}

\item{resolution}{Resolution of the output raster (in units relevant to `coordinate.system`). Only required if prediction.grid is irregular, and thus needs to be rasterised. Defaults to NULL.}
}
\value{
A list object with two elements, namely values of the extrapolation metrics and their spatial representation in raster form
}
\description{
Assesses univariate (Type I) and combinatorial (Type II) extrapolation in density surface models of line transect data. Models are built in a reference (calibration) system and projected into one or more target (prediction) systems. The function is based on original code from the \code{\href{https://cran.r-project.org/web/packages/ecospat/index.html}{ecospat}} package, and tailored to outputs returned by the \code{\href{https://cran.r-project.org/web/packages/dsm/index.html}{dsm}} package, although it can be applied to other predictive modelling scenarios. See Bouchet et al. (2019) and Mesgaran et al. (2014) for details.
}
\details{
The function calculates values of the ExDet (EXtrapolation DETection) metric as originally proposed by Mesgaran et al. (2014). ExDet is strictly negative during univariate extrapolation (i.e. when predictions are made outside the range of individual covariates), strictly positive and >1 during combinatorial extrapolation (i.e. when predictions are made within the range of individual covariates, but for combinations of envionmental conditions not encountered in the sample), and lies within the range 0-1 when predictions are made in conditions analogous to those found in the reference system. The function also determines which covariate makes the largest contribution to each type of extrapolation; this is the most influential covariate (MIC). See Mesgaran et al. (2014) for details.

The `data` list captures ExDet values in prediction locations (i.e. cells in prediction.grid) and is organised into multiple data.frame objects, as follows:

\tabular{ll}{
  \code{all} \tab All prediction locations\cr
  \code{univariate} \tab Prediction locations subject to univariate extrapolation (only)\cr
  \code{combinatorial} \tab Prediction locations subject to combinatorial extrapolation (only)\cr
  \code{analogue} \tab Prediction locations where conditions are analogous to sampled conditions (only)\cr
 }

Each data.frame contains four columns:

\tabular{ll}{
  \code{ExDet} \tab ExDet values\cr
  \code{mic_univariate} \tab Integer identifying the univariate MIC\cr
  \code{mic_combinatorial} \tab Integer identifying the combinatorial MIC\cr
  \code{mic} \tab Integer identifying the MIC()\cr
 }

The `rasters` list consists of two elements, named `ExDet` and `mic`. Each contains individual rasters mapping ExDet values and MICs, respectively.
}
\references{
Bouchet PJ, Miller DL, Roberts JJ, Mannocci L, Harris CM and Thomas L (2019). From here and now to there and then: Practical recommendations for extrapolating cetacean density surface models to novel conditions. CREEM Technical Report 2019-01, 59 p. \href{https://research-repository.st-andrews.ac.uk/handle/10023/18509}{https://research-repository.st-andrews.ac.uk/handle/10023/18509}

Mesgaran MB, Cousens RD, Webber BL (2014). Here be dragons: a tool for quantifying novelty due to covariate range and correlation change when projecting species distribution models. Diversity & Distributions, 20: 1147-1159, DOI: \href{https://onlinelibrary.wiley.com/doi/full/10.1111/ddi.12209}{10.1111/ddi.12209}
}
\author{
Phil J. Bouchet
}
