% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmextra-package.R
\docType{package}
\name{dsmextra}
\alias{dsmextra}
\alias{dsmextra-package}
\title{Extrapolation detection for density surface models}
\description{
The dsmextra package provides functions for detecting, diagnosing and visualising extrapolation in multivariate environmental space, with applications to density surface models (DSMs) created in the \link{dsm-package} package.
}
