% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExDet.R
\name{ExDet}
\alias{ExDet}
\title{EXtrapolation DETection tool}
\usage{
ExDet(ref, tg, xp)
}
\arguments{
\item{ref}{Reference data. A data.frame with the values of the variables (i.e. columns) for each sample unit (e.g. segments).}

\item{tg}{Target data. A data.frame with the values of the variables (i.e. columns) for each point of the prediction extent.}

\item{xp}{Character string. Names of the covariates of interest.}
}
\value{
Returns a tibble with four columns. (1) \code{ExDet}: Value of the ExDet metric (negative for univariate extrapolation, >1 for combinatorial extrapolation, within the range 0-1 for analogue conditions). (2) \code{mic_univariate}: Integer indicating the covariate with the largest contribution to univariate extrapolation (most influential covariate, MIC).  (3) \code{mic_combinatorial}: Integer indicating the covariate with the largest contribution to combinatorial extrapolation. (4) \code{mic}: most influential covariate.
}
\description{
Assesses univariate (Type I) and combinatorial (Type II) extrapolation between a reference system (ref) and a projection system (p). See Mesgaran et al. (2014) for an explanation. This function is an updated version of some original code from the \code{\link[ecospat]{ecospat}} package (Broennimann et al. 2016).
}
\references{
Mesgaran, M.B., Cousens, R.D. & Webber, B.L. (2014) \href{https://onlinelibrary.wiley.com/doi/full/10.1111/ddi.12209}{Here be dragons: a tool for quantifying novelty due to covariate range and correlation change when projecting species distribution models}. Diversity & Distributions, 20: 1147-1159, DOI: 10.1111/ddi.12209

Broennimann O, Di Cola V, Guisan A (2016). ecospat: Spatial Ecology Miscellaneous Methods. R package version 2.1.1. \href{https://CRAN.R-project.org/package=ecospat}{https://CRAN.R-project.org/package=ecospat}.
}
\author{
Phil J. Bouchet
}
\keyword{internal}
