% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_ras}
\alias{mean_ras}
\title{---------------------------------------------
---------------------------------------------}
\usage{
mean_ras(x, ...)
}
\description{
---------------------------------------------
---------------------------------------------
}
