% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_and_p}
\alias{n_and_p}
\title{---------------------------------------------
---------------------------------------------}
\usage{
n_and_p(x)
}
\description{
---------------------------------------------
---------------------------------------------
}
