% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.extrapolation_results_summary.R
\name{print.extrapolation_results_summary}
\alias{print.extrapolation_results_summary}
\title{Print extrapolation results summary}
\usage{
\method{print}{extrapolation_results_summary}(x, digits = 2, ...)
}
\arguments{
\item{x}{\code{summary} element of the result of \code{\link{compute_extrapolation}}}

\item{digits}{precision of results}

\item{\dots}{for S3 compatability}
}
\value{
invisibly returns the summary part of the object only, printing the results
}
\description{
Print the summary of extrapolation results (already calculated by \code{\link{compute_extrapolation}}).
}
\author{
David L Miller
}
\keyword{internal}
