% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale_cov}
\alias{rescale_cov}
\title{---------------------------------------------
---------------------------------------------}
\usage{
rescale_cov(ynew, y)
}
\description{
---------------------------------------------
---------------------------------------------
}
