% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_raster}
\alias{safe_raster}
\title{---------------------------------------------
---------------------------------------------}
\usage{
safe_raster(...)
}
\description{
---------------------------------------------
---------------------------------------------
}
