% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_extrapolation.R
\name{summarise_extrapolation}
\alias{summarise_extrapolation}
\title{Summary of extrapolation}
\usage{
summarise_extrapolation(
  extrapolation.object,
  covariate.names = NULL,
  extrapolation = TRUE,
  mic = TRUE
)
}
\arguments{
\item{extrapolation.object}{Output object from a run of \link{compute_extrapolation}.}

\item{covariate.names}{Character string. Names of the covariates of interest.}

\item{extrapolation}{Logical. Whether to return a summary of univariate/combinatorial extrapolation. Defaults to TRUE.}

\item{mic}{Logical. Whether to return a summary of the most influential covariates (MIC) - see \link{compute_extrapolation}. Defaults to TRUE.}
}
\value{
Prints a summary table in the R console. In addition, if assigned to an object, returns a list with the table values (.n = number of locations, .p = corresponding percentage).
}
\description{
Displays a tabular summary of the geographic extent of extrapolation. This is calculated as the number (and proportion) of prediction locations (i.e. grid cells) subject to extrapolation.
}
\examples{
library(dsmextra)

# Load the Mid-Atlantic sperm whale data (see ?spermwhales)
data(spermwhales)

# Extract the data
segs <- spermwhales$segs
predgrid <- spermwhales$predgrid

# Define relevant coordinate system
my_crs <- sp::CRS("+proj=aea +lat_1=38 +lat_2=30 +lat_0=34 +lon_0=-73 +x_0=0
 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0")

# Assess extrapolation in the multivariate space defined by five covariates
spermw.extrapolation <- compute_extrapolation(segments = segs,
      covariate.names = c("Depth", "DistToCAS", "SST", "EKE", "NPP"),
      prediction.grid = predgrid,
      coordinate.system = my_crs)

# Summarise extrapolation
spermw.summary <- summarise_extrapolation(extrapolation.object = spermw.extrapolation,
     covariate.names = c("Depth", "DistToCAS", "SST", "EKE", "NPP"),
     extrapolation = TRUE,
     mic = TRUE)

print(spermw.summary)
}
\references{
Bouchet PJ, Miller DL, Roberts JJ, Mannocci L, Harris CM and Thomas L (2019). From here and now to there and then: Practical recommendations for extrapolating cetacean density surface models to novel conditions. CREEM Technical Report 2019-01, 59 p. \href{https://research-repository.st-andrews.ac.uk/handle/10023/18509}{https://research-repository.st-andrews.ac.uk/handle/10023/18509}

Mesgaran MB, Cousens RD, Webber BL (2014). Here be dragons: a tool for quantifying novelty due to covariate range and correlation change when projecting species distribution models. Diversity & Distributions, 20: 1147-1159, DOI: \href{https://onlinelibrary.wiley.com/doi/full/10.1111/ddi.12209}{10.1111/ddi.12209}
}
\seealso{
\code{\link{compute_extrapolation}}
}
\author{
Phil J. Bouchet
}
