% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.extrapolation_results.R
\name{summary.extrapolation_results}
\alias{summary.extrapolation_results}
\title{Summarize extrapolation results}
\usage{
\method{summary}{extrapolation_results}(object, digits, ...)
}
\arguments{
\item{object}{result of running \code{\link{compute_extrapolation}}}

\item{digits}{precision of results}

\item{\dots}{for S3 compatability}
}
\value{
invisibly returns the summary part of the object only, printing the results
}
\description{
Print the summary of extrapolation results (already calculated by \code{\link{compute_extrapolation}}).
}
\author{
David L Miller
}
\keyword{internal}
