% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatif.opt.R
\name{whatif.opt}
\alias{whatif.opt}
\title{Counterfactual evaluation, optimised for dsmextra}
\usage{
whatif.opt(
  formula = NULL,
  data,
  cfact,
  nearby = 1,
  miss = "list",
  no.partitions
)
}
\arguments{
\item{formula}{An optional formula without a dependent variable, allowing transformations of combinations of the variables in both data and cfact.}

\item{data}{Either a model output object, or a n-by-k non-character (logical or numeric) matrix or data frame of observed covariate data with n data points or units and k covariates.}

\item{cfact}{Counterfactuals.}

\item{nearby}{Scalar indicating which reference data points are considered to be 'nearby' (i.e. within ‘nearby’ mean geometric Gower's distance of) prediction points.}

\item{miss}{Optional string indicating the strategy for dealing with missing data.}

\item{no.partitions}{Integer. Number of desired partitions of the data (default of 10).}
}
\value{
A list object containing extrapolation values in both data.frame and raster format.
}
\description{
Implements the methods described in King and Zeng (2007) for evaluating counterfactuals.
}
\details{
The \code{\link[WhatIf]{whatif}} function from the \href{https://CRAN.R-project.org/package=WhatIf}{Whatif} package (Gandrud et al. 2017) may not run on very large datasets. To circumvent this problem, \code{whatif.opt} sets the calculations performed by \code{\link[WhatIf]{whatif}} to run on partitions of the data instead, for greater efficiency. \code{whatif.opt} can be called internally within compute_nearby by using two additional arguments, namely:
\tabular{ll}{
  \code{max.size} \tab Threshold above which partitioning will be triggered \cr
  \code{no.partitions} \tab Number of required partitions \cr
 }
 In practice, a run of \code{\link{compute_nearby}} begins with a quick assessment of the dimensions of the input data, i.e. the reference and target data.frames. If the product of their dimensions (i.e. number of segments multiplied by number of prediction grid cells) exceeds the value set for \code{max.size}, then \code{no.partitions} subsets of the data will be created and the computations run on each using map functions from the \code{\link{purrr}} package (Henry and Wickham 2019). This means that a smaller \code{max.size} will trigger partitioning on correspondingly smaller datasets. By default, \code{max.size} is set to \code{1e7}. This value was chosen arbitrarily, and should be sufficiently large as to obviate the need for partitioning on most datasets.
}
\references{
Gandrud C, King G, Stoll H, Zeng L (2017). WhatIf: Evaluate Counterfactuals. R package version 1.5-9. \href{https://CRAN.R-project.org/package=WhatIf}{https://CRAN.R-project.org/package=WhatIf}.

King G, Zeng L (2007). When can history be our guide? The pitfalls of counterfactual inference. International Studies Quarterly 51, 183–210. DOI: \href{https://www.jstor.org/stable/pdf/4621707.pdf?seq=1#page_scan_tab_contents}{10.1111/j.1468-2478.2007.00445.x}

Henry L, Wickham H (2019). purrr: Functional Programming Tools. R package version 0.3.2. \href{https://CRAN.R-project.org/package=purrr}{https://CRAN.R-project.org/package=purrr}.
}
\seealso{
\code{\link{compute_nearby}}, \code{\link[WhatIf]{whatif}}
}
\author{
Phil J. Bouchet
}
