% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_covariates.R
\name{compare_covariates}
\alias{compare_covariates}
\title{Compare different combinations of covariates}
\usage{
compare_covariates(extrapolation.type = "both", segments,
  covariate.names, n.covariates = NULL, prediction.grid,
  coordinate.system, create.plots = TRUE, display.percent = TRUE)
}
\arguments{
\item{extrapolation.type}{Character string. Type of extrapolation to be assessed. Can be one of \code{univariate}, \code{combinatorial}, or \code{both} (default).}

\item{segments}{Segment data.frame (i.e. surveyed transects divided into segments for analysis). This is the reference dataset used for model building and calibration. This must contain one column for each of the covariates in \code{covariate.names}.}

\item{covariate.names}{Character string. Names of the covariates of interest.}

\item{n.covariates}{Maximum number of covariates. The function will compare all combinations of 1 to \code{n.covariates} covariates.}

\item{prediction.grid}{Prediction data.frame. This contains both geographic coordinates (\code{x}, \code{y}) and covariate values associated with the target locations for which predictions are desired. Typically, these locations are taken as the centroids of the grid cells in a spatial prediction grid/raster.}

\item{coordinate.system}{Projected coordinate system relevant to the study location. Can be either a character string or an object of class \code{\link[sp]{CRS}}.}

\item{create.plots}{Logical, defaults to \code{TRUE}. Whether to produce summary plots.}

\item{display.percent}{Logical. If \code{TRUE} (default), scales the y-axis of the summary plots as a percentage of the total number of grid cells in \code{prediction.grid}.}
}
\value{
Prints a summary table in the R console. Also generates summary boxplots if \code{create.plots} is set to \code{TRUE}.
}
\description{
Summarises the extent of  univariate (Type I) and combinatorial (Type II) extrapolation associated with different combinations of input covariates.
}
\details{
The extent and magnitude of extrapolation naturally vary with the type and number of covariates considered. It may be useful, therefore, to test different combinations of covariates to inform their selection \emph{a priori}, i.e. before model fitting, thereby supporting model parsimony.
}
\examples{
library(dsmextra)

# Load the Mid-Atlantic sperm whale data (see ?spermwhales)
data(spermwhales)

# Extract the data
segs <- spermwhales$segs
predgrid <- spermwhales$predgrid

# Define relevant coordinate system
my_crs <- sp::CRS("+proj=aea +lat_1=38 +lat_2=30 +lat_0=34 +lon_0=-73 +x_0=0
 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0")

# Define covariates
my_cov <- c("Depth", "SST", "NPP", "DistToCAS", "EKE")

# Compare the extent of univariate and combinatorial
# extrapolation for all combinations of 1 to 5 covariates
compare_covariates(extrapolation.type = "both",
                  covariate.names = my_cov,
                  n.covariates = NULL,
                  segments = segs,
                  prediction.grid = predgrid,
                  coordinate.system = my_crs,
                  create.plots = TRUE,
                  display.percent = TRUE)
}
\references{
Bouchet PJ, Miller DL, Roberts JJ, Mannocci L, Harris CM and Thomas L (2019). From here and now to there and then: Practical recommendations for extrapolating cetacean density surface models to novel conditions. CREEM Technical Report 2019-01, 59 p. \href{https://research-repository.st-andrews.ac.uk/handle/10023/18509}{https://research-repository.st-andrews.ac.uk/handle/10023/18509}

Mesgaran MB, Cousens RD, Webber BL (2014). Here be dragons: a tool for quantifying novelty due to covariate range and correlation change when projecting species distribution models. Diversity & Distributions, 20: 1147-1159, DOI: \href{https://onlinelibrary.wiley.com/doi/full/10.1111/ddi.12209}{10.1111/ddi.12209}
}
\seealso{
\code{\link{compute_extrapolation}}, \code{\link{summarise_extrapolation}}
}
\author{
Phil J. Bouchet
}
