% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmextra-package.R
\docType{package}
\name{dsmextra}
\alias{dsmextra}
\title{Extrapolation detection in density surface models}
\description{
The \code{dsmextra} package provides functions for detecting, diagnosing and visualising extrapolation in multivariate environmental space, with applications to density surface models (DSMs) created in the \code{\link[dsm]{dsm}} package.
}
\details{
Further information on distance sampling methods and example code is available at \href{http://distancesampling.org/R/}{http://distancesampling.org/R/}.

For help with distance sampling, there is a Google Group \href{https://groups.google.com/forum/#!forum/distance-sampling}{https://groups.google.com/forum/#!forum/distance-sampling}.
}
\author{
Phil J. Bouchet, Laura Mannocci, David Miller
}
