% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmextra-package.R
\docType{data}
\name{spermwhales}
\alias{spermwhales}
\title{Sperm whale sightings in the Mid-Atlantic}
\format{
A list of two
\describe{
  \item{segs}{segment-level data, as used in the \code{\link[dsm]{dsm}} package}
  \item{predgrid}{prediction data.frame}
}
}
\source{
The data are held and described on OBIS-SEAMAP at: \url{http://seamap.env.duke.edu/dataset/396}.
}
\description{
Data from a combination of several NOAA visual surveys conducted for cetaceans in the U.S. Mid-Atlantic. Survey tracklines were split into segments for analysis in the \code{\link[dsm]{dsm}} R package. Spatial coordinates (\code{x}, \code{y}) and values of both static and dynamic environmental covariates are availble for each segment mid-point and for the centroids of a 10 km grid overlaid on the study area. Covariates include: depth, sea surface temperature (\code{SST}), net primary production (\code{NPP}), distance to nearest canyon or seamount (\code{DistToCAS}), and eddy kinetic energy (\code{EKE}).
}
\details{
An example extrapolation assessment is provided at \url{https://densitymodelling.github.io/model-extrapolation/vignette/Extrapolation-vignette.html}.
}
\references{
NOAA Northeast Fisheries Science Center (2004). A survey for abundance and distribution of cetaceans in the U.S. Mid-Atlantic with an emphasis on pilot whales. Report of a survey conducted aboard NOAA Ship Gordon Gunter Cruise GU-04-03 (028).

NOAA Northeast Fisheries Science Center (2004). Report on the 2004 Mid-Atlantic Marine Mammal Shipboard Abundance Survey aboard the R/V Endeavor, Cruise No. EN 04-395/396.
}
\keyword{datasets}
