% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_extrapolation.R
\name{summarise_extrapolation}
\alias{summarise_extrapolation}
\title{Summary of extrapolation}
\usage{
summarise_extrapolation(
  extrapolation.object,
  covariate.names = NULL,
  extrapolation = TRUE,
  mic = TRUE
)
}
\arguments{
\item{extrapolation.object}{Output object from a run of \link{compute_extrapolation}.}

\item{extrapolation}{Logical. Whether to return a summary of univariate/combinatorial extrapolation. Defaults to TRUE.}

\item{mic}{Logical. Whether to return a summary of the most influential covariates (MIC) - see \link{compute_extrapolation}. Defaults to TRUE.}
}
\value{
Prints a summary table in the R console. In addition, if assigned to an object, returns a list with the table values (.n = number of locations, .p = corresponding percentage).
}
\description{
Displays a tabular summary of the geographic extent of extrapolation. This is calculated as the number (and proportion) of prediction locations (i.e. grid cells) subject to extrapolation.
}
\references{
Bouchet PJ, Miller DL, Roberts JJ, Mannocci L, Harris CM and Thomas L (2019). From here and now to there and then: Practical recommendations for extrapolating cetacean density surface models to novel conditions. CREEM Technical Report 2019-01, 59 p. \href{https://research-repository.st-andrews.ac.uk/handle/10023/18509}{https://research-repository.st-andrews.ac.uk/handle/10023/18509}

Mesgaran MB, Cousens RD, Webber BL (2014). Here be dragons: a tool for quantifying novelty due to covariate range and correlation change when projecting species distribution models. Diversity & Distributions, 20: 1147-1159, DOI: \href{https://onlinelibrary.wiley.com/doi/full/10.1111/ddi.12209}{10.1111/ddi.12209}
}
\seealso{
\code{\link{compute_extrapolation}}
}
\author{
Phil J. Bouchet
}
\keyword{internal}
